Adding share links and add article button for admin

+++ db/migrate/20191102173407_create_posts.rb
@@ -3,7 +3,7 @@ class CreatePosts < ActiveRecord::Migration[5.2]
     create_table :posts do |t|
       t.string :title
       t.text :content
-      t.numeric :like
+      t.integer :like
       t.belongs_to :user, index: true
 
       t.timestamps
