+++ db/migrate/20191102173453_create_comments.rb
@@ -2,7 +2,7 @@ class CreateComments < ActiveRecord::Migration[5.2]
   def change
     create_table :comments do |t|
       t.text :content
-      t.numeric :like
+      t.integer :like
       t.belongs_to :user, index: true
       t.belongs_to :post, index: true
